+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - add layer action to databrowser
 - add optional default CSS
 - allow to close panel when editing by ctrl-Enter when in textarea
+- add management for map max bounds
 
 ## 0.4.x
 - add a data browser